Déchiffrement et plus aucun warning

+++ include/argumentsParseur.h
@@ -9,8 +9,8 @@ class ArgumentsParser
 public:
     ArgumentsParser(std::vector<std::string> nomsOptions);
     void parse(int argc, char* argv[]);
-    std::string getArgument(int index);
-    int nbArguments(){return arguments.size();};
+    std::string getArgument(size_t index);
+    size_t nbArguments(){return arguments.size();};
     bool getOption(std::string nom);
     bool demandeAide();
     static std::vector<std::string> getNomsOptions();
